+++ C21-Ex01-ItamarAyalon-206024796-YanivAvrahami-302374921/UI/Factory Method/FormFactory.cs
@@ -19,7 +19,7 @@ namespace UI
 
             if (i_FormType.Equals(typeof(AlbumsForm)))
             {
-                newForm = new AlbumsForm();
+                newForm = new AlbumsForm(this);
             }
             else if (i_FormType.Equals(typeof(AlbumForm)))
             {
